Separated services into its own component, DRYed its code and removed unused imports on app/page.js
