+++ app/page.js
@@ -1,22 +1,14 @@
 import "./HomePage.css";
 
-import { Container, Row, Col, Card } from "react-bootstrap";
+import { Container, Row, Col } from "react-bootstrap";
 import Carousel from "react-bootstrap/Carousel";
 import CarouselItem from "react-bootstrap/CarouselItem";
 import CarouselCaption from "react-bootstrap/CarouselCaption";
-import {
-  FaQuoteLeft,
-  FaMobile,
-  FaCogs,
-  FaProjectDiagram,
-  FaShoppingCart,
-  FaLaptopCode,
-} from "react-icons/fa";
+import { FaQuoteLeft } from "react-icons/fa";
 import { LuTrendingUp, LuUsers, LuLightbulb } from "react-icons/lu";
 
 import HomeHero from "@/components/HomeHero";
 import BlogCard from "@/components/BlogCard";
-import BlackCard from "@/components/BlackCard";
 import CustomButton from "@/components/CustomButton";
 import CallToAction from "@/components/CallToAction";
 
@@ -24,7 +16,7 @@ import { fetchBlogs } from "@/services/blogService";
 import { fetchClients } from "@/services/clientService";
 import defaultClientImage from "@/public/assets/homepage/default-pic.jpg";
 
-import ServiceCards from "@/components/ServiceCards";
+import Services from "@/components/Services";
 import BlackCardsSection from "@/components/BlackCardsSection";
 import AOSWrapper from "@/components/AOSWrapper";
 import Image from "next/image";
@@ -32,42 +24,6 @@ import Image from "next/image";
 export default async function HomePage() {
   const blogs = await fetchBlogs();
   const clients = await fetchClients();
-
-  const services = [
-    {
-      image: "/assets/homepage/services-img/market.webp",
-      icon: <FaMobile size={40} />,
-      title: "Marketing + Social Media",
-      description:
-        "Amplify your brand's impact with expert marketing and social media strategies. Engage your audience, accelerate growth, and redefine your digital presence.",
-    },
-    {
-      image: "/assets/homepage/services-img/ecom.webp",
-      icon: <FaShoppingCart size={40} />,
-      title: "E-commerce Solutions",
-      description:
-        "Transform your online business with tailored e-commerce solutions. Optimize performance, boost sales, and enhance customer satisfaction.",
-    },
-    {
-      image: "/assets/homepage/services-img/business.webp",
-      icon: <FaCogs size={40} />,
-      title: "Business Process Optimization",
-      description: "Streamline your operations and enhance efficiency with our business optimization services. Achieve greater productivity and profitability.",
-    },
-    {
-      image: "/assets/homepage/services-img/manage.webp",
-      icon: <FaProjectDiagram size={40} />,
-      title: "Project Management",
-      description:
-        "Ensure timely, budget-friendly project delivery with our expert management services. We focus on seamless project process management with an emphasis on continuity to improve efficiency and achieve your goals.",
-    },
-    {
-      image: "/assets/homepage/services-img/it.webp",
-      icon: <FaLaptopCode size={40} />,
-      title: "IT + Tech Integrations",
-      description: "Boost your business with our innovative IT solutions. We design custom IT and tech packages to increase efficiencies and maximize output.",
-    },
-  ];
 
   const reasons = [
     {
@@ -189,7 +145,7 @@ export default async function HomePage() {
               </Row>
             </Col>
             <Col md={12} lg={7}>
-              <ServiceCards services={services} />
+              <Services />
             </Col>
           </Row>
         </Container>
@@ -217,7 +173,7 @@ export default async function HomePage() {
                     alt={client.name}
                     width={100}
                     height={100}
-                    // style={{ borderRadius: "50%" }}
+                  // style={{ borderRadius: "50%" }}
                   />
                   <p className="mb-4 fs-5">
                     <FaQuoteLeft className="quote-icon" size={25} />
